requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
pathlib>=1.0.1 